boot: Add DTB file and bootloader support for RPi 5
* bcm2712-rpi-5-b.dtb required
* Bootloader uses ARM PL011 UART on RPi 5 (via UART connector)

+++ tools/bootloader/bootloader07.c
@@ -8,10 +8,10 @@
 // 8  TX out
 // 10 RX in
 
-extern void PUT32 ( unsigned int, unsigned int );
-extern void PUT16 ( unsigned int, unsigned int );
-extern void PUT8 ( unsigned int, unsigned int );
-extern unsigned int GET32 ( unsigned int );
+extern void PUT32 ( unsigned long, unsigned int );
+extern void PUT16 ( unsigned long, unsigned int );
+extern void PUT8 ( unsigned long, unsigned int );
+extern unsigned int GET32 ( unsigned long );
 extern unsigned int GETPC ( void );
 extern void BRANCHTO ( unsigned int );
 extern void dummy ( unsigned int );
